+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 6. `FormeController`'s data renamed to value
 7. remove readOnly getter & setter from `FormeController`
 8. remove `mounted` from `FormeFieldController` , use `isDisposed` instead
+9. when `FormeFieldController` or `FormeController ` is referenced  by others , `FormeFieldState`&`_FormeState` can recycled after disposed
+10. remove `compareValue` from `FormeFieldState`
 
 ## 3.5.2